In the drag_color.py script, the patch attribute is called from Colorbar. However, in more recent versions of Matplotlib (starting from version 3.4.0), the Colorbar object no longer has a patch attribute. Instead, it uses the outline attribute to access the rectangular patch of the colorbar. Therefore, replacing all instances of patch with outline in the script should fix this issue.
This issue occurs because the Colorbar object was updated in recent versions of Matplotlib, and some of the attributes and methods have been changed or removed. As a result, code that was written for older versions of Matplotlib may not work correctly in newer versions.

Note: If you have a version of tayph that does work you may need to updated your version of matplotlib.